+++ pydidas/core/constants/generic_params_data_import.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -19,8 +21,8 @@ Parameters for data import.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
-__license__ = "GPL-3.0"
+__copyright__ = "Copyright 2021- Helmholtz-Zentrum Hereon"
+__license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"
 __all__ = ["GENERIC_PARAMS_DATA_IMPORT"]
@@ -174,6 +176,15 @@ GENERIC_PARAMS_DATA_IMPORT = {
             "the number of frames in the hdf5 dataset. A value -1 auto-discovers "
             "the number of images per file."
         ),
+    },
+    "_counted_images_per_file": {
+        "type": int,
+        "default": 1,
+        "name": "Processed images per file",
+        "choices": None,
+        "unit": "",
+        "allow_None": False,
+        "tooltip": "The number of images per file pydidas counted from the first file.",
     },
     "directory_path": {
         "type": "Path",